GPS usage introduced

+++ app/src/main/java/com/ibashniak/weatherapp/network/dto/Wind.kt
@@ -1,6 +1,6 @@
 package com.ibashniak.weatherapp.network.dto
 
 data class Wind(
-    val deg: Int,
+    val deg: Double,
     val speed: Double
 )
